Created List component and updated Page
